HealthcareTranslationApp/requirements.txt
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-core==2.4.2
google-cloud-speech==2.31.0
google-cloud-texttospeech==2.25.0
google-cloud-translate==3.20.1
googleapis-common-protos==1.68.0
grpc-google-iam-v1==0.14.0
grpcio==1.70.0
grpcio-status==1.70.0
idna==3.10
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydub==0.25.1
requests==2.32.3
rsa==4.9
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
